Refactors the project.
