+++ src/EvalutionCS/Ast/AstBuilder.cs
@@ -124,294 +124,4 @@ namespace EvalutionCS.Ast
             return expressions;
         }
     }
-
-    public abstract class Expression
-    {
-    }
-
-    public class BinaryExpression : Expression
-    {
-        public BinaryExpression(Expression leftExpression, BinaryOperator binaryOperator, Expression rightExpression)
-        {
-            LeftExpression = leftExpression;
-            BinaryOperator = binaryOperator;
-            RightExpression = rightExpression;
-        }
-
-        public Expression LeftExpression { get; set; }
-        public BinaryOperator BinaryOperator { get; set; }
-        public Expression RightExpression { get; set; }
-
-        public override bool Equals(object obj)
-        {
-            if (obj is BinaryExpression)
-            {
-                var typedObj = obj as BinaryExpression;
-                return typedObj.LeftExpression.Equals(LeftExpression)
-                       && typedObj.BinaryOperator.Equals(BinaryOperator)
-                       && typedObj.RightExpression.Equals(RightExpression);
-            }
-            return false;
-        }
-
-    }
-
-    public class LiteralExpression : Expression
-    {
-        public LiteralExpression(Literal literal)
-        {
-            Literal = literal;
-        }
-
-        public Literal Literal { get; set; }
-
-        public override bool Equals(object obj)
-        {
-            if (obj is LiteralExpression)
-            {
-                var typedObj = obj as LiteralExpression;
-                return typedObj.Literal.Equals(Literal);
-            }
-            return false;
-        }
-    }
-    public class UnaryExpression : Expression
-    {
-        public UnaryExpression(UnaryOperator unaryOperator, Expression expression)
-        {
-            UnaryOperator = unaryOperator;
-            Expression = expression;
-        }
-
-        public UnaryOperator UnaryOperator { get; set; }
-        public Expression Expression { get; set; }
-
-        public override bool Equals(object obj)
-        {
-            if (obj is UnaryExpression)
-            {
-                var typedObj = obj as UnaryExpression;
-                return typedObj.UnaryOperator.Equals(UnaryOperator)
-                    && typedObj.Expression.Equals(Expression);
-            }
-            return false;
-        }
-
-    }
-    public class MultiCallExpression : Expression
-    {
-        public MultiCallExpression(Multicall multicall)
-        {
-            Multicall = multicall;
-        }
-
-        public Multicall Multicall { get; set; }
-
-        public override bool Equals(object obj)
-        {
-            if (obj is MultiCallExpression)
-            {
-                var typedObj = obj as MultiCallExpression;
-                return typedObj.Multicall.Equals(Multicall);
-            }
-            return false;
-        }
-
-    }
-
-    public abstract class Literal
-    {
-        
-    }
-
-    public class BoolLiteral : Literal
-    {
-        public bool Value { get; set; }
-        public override bool Equals(object obj)
-        {
-            if (obj is BoolLiteral)
-            {
-                return (obj as BoolLiteral).Value.Equals(Value);
-            }
-            return false;
-        }
-    }
-    public class Int32Literal : Literal
-    {
-        public Int32Literal(int value)
-        {
-            Value = value;
-        }
-
-        public int Value { get; private set; }
-
-        public override bool Equals(object obj)
-        {
-            if (obj is Int32Literal)
-            {
-                return (obj as Int32Literal).Value.Equals(Value);
-            }
-            return false;
-        }
-    }
-    public class DoubleLiteral : Literal
-    {
-        public DoubleLiteral(double value)
-        {
-            Value = value;
-        }
-
-        public double Value { get; set; }
-
-        public override bool Equals(object obj)
-        {
-            if (obj is DoubleLiteral)
-            {
-                return (obj as DoubleLiteral).Value.Equals(Value);
-            }
-            return false;
-        }
-    }
-
-    public abstract class Multicall
-    {
-    }
-
-    public class CurrentContextMethodCall : Multicall
-    {
-        public CurrentContextMethodCall(string identifier, List<Expression> arguments)
-        {
-            Identifier = identifier;
-            Arguments = arguments;
-        }
-
-        public string Identifier { get; set; }
-        public List<Expression> Arguments { get; set; }
-
-        public override bool Equals(object obj)
-        {
-            if (obj is CurrentContextMethodCall)
-            {
-                var typedObj = obj as CurrentContextMethodCall;
-                for (var i = 0; i < Arguments.Count; i++)
-                {
-                    if (!Arguments[i].Equals(typedObj.Arguments[i]))
-                    {
-                        return false;
-                    }
-                }
-                return typedObj.Identifier.Equals(Identifier);
-            }
-            return false;
-        }
-
-    }
-    public class CurrentContextPropertyCall : Multicall
-    {
-        public CurrentContextPropertyCall(string identifier)
-        {
-            Identifier = identifier;
-        }
-
-        public string Identifier { get; set; }
-
-        public override bool Equals(object obj)
-        {
-            if (obj is CurrentContextPropertyCall)
-            {
-                return (obj as CurrentContextPropertyCall).Identifier.Equals(Identifier);
-            }
-            return false;
-        }
-
-    }
-    public class ObjectContextPropertyCall : Multicall
-    {
-        public ObjectContextPropertyCall(Multicall multicall, string identifier)
-        {
-            Multicall = multicall;
-            Identifier = identifier;
-        }
-
-        public Multicall Multicall { get; set; }
-        public string Identifier { get; set; }
-
-        public override bool Equals(object obj)
-        {
-            if (obj is ObjectContextPropertyCall)
-            {
-                return (obj as ObjectContextPropertyCall).Identifier.Equals(Identifier)
-                       && (obj as ObjectContextPropertyCall).Multicall.Equals(Multicall);
-            }
-            return false;
-        }
-
-    }
-    public class ObjectContextMethodCall : Multicall
-    {
-        public ObjectContextMethodCall(Multicall multicall, string identifier, List<Expression> arguments)
-        {
-            Multicall = multicall;
-            Identifier = identifier;
-            Arguments = arguments;
-        }
-
-        public Multicall Multicall { get; set; } // rename to PrevCall
-        public string Identifier { get; set; }
-        public List<Expression> Arguments { get; set; }
-
-        public override bool Equals(object obj)
-        {
-            if (obj is ObjectContextMethodCall)
-            {
-                var typedObj = obj as ObjectContextMethodCall;
-                for (var i = 0; i < Arguments.Count; i++)
-                {
-                    if (!Arguments[i].Equals(typedObj.Arguments[i]))
-                    {
-                        return false;
-                    }
-                }
-                return typedObj.Identifier.Equals(Identifier)
-                    && typedObj.Multicall.Equals(Multicall);
-            }
-            return false;
-        }
-
-    }
-    public class ArrayElementCall : Multicall
-    {
-        public ArrayElementCall(Multicall multicall, Expression expression)
-        {
-            Multicall = multicall;
-            Expression = expression;
-        }
-
-        public Multicall Multicall { get; set; }
-        public Expression Expression { get; set; }
-
-        public override bool Equals(object obj)
-        {
-            if (obj is ArrayElementCall)
-            {
-                return (obj as ArrayElementCall).Expression.Equals(Expression)
-                       && (obj as ArrayElementCall).Multicall.Equals(Multicall);
-            }
-            return false;
-        }
-
-    }
-    public enum BinaryOperator
-    {
-        Add,
-        Subtract,
-        Multiply,
-        Divide
-    }
-    public enum UnaryOperator
-    {
-        LogicalNegate,
-        Negate,
-        Identity,
-    }
 }
