+++ src/EvalutionCS/Ast/BinaryOperator.cs
@@ -1,0 +1,10 @@
+namespace EvalutionCS.Ast
+{
+    public enum BinaryOperator
+    {
+        Add,
+        Subtract,
+        Multiply,
+        Divide
+    }
+}
