+++ src/EvalutionCS/Ast/Multicall.cs
@@ -1,0 +1,10 @@
+namespace EvalutionCS.Ast
+{
+    using System;
+
+    public abstract class Multicall
+    {
+        public abstract Type BuildBody(BuildArguments args);
+        public abstract Type GetExpressionType(BuildArguments args);
+    }
+}
