+++ src/EvalutionCS/Ast/UnaryOperator.cs
@@ -1,0 +1,9 @@
+namespace EvalutionCS.Ast
+{
+    public enum UnaryOperator
+    {
+        LogicalNegate,
+        Negate,
+        Identity,
+    }
+}
